+++ services/core/java/core-flow-service/src/main/java/org/ubiquia/core/flow/service/cluster/FlowClusterService.java
@@ -2,6 +2,7 @@ package org.ubiquia.core.flow.service.cluster;
 
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
+import java.util.Objects;
 import java.util.concurrent.ThreadLocalRandom;
 import org.jgroups.JChannel;
 import org.jgroups.Message;
@@ -9,6 +10,7 @@ import org.jgroups.Receiver;
 import org.jgroups.View;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -41,8 +43,14 @@ public class FlowClusterService implements Receiver {
     @Value("${ubiquia.cluster.port}")
     private int port;
 
+    @Value("${ubiquia.cluster.bind-addr:GLOBAL}")
+    private String bindAddr;
+
     @Value("${ubiquia.cluster.rejoin-delay-milliseconds:5000}")
     private long rejoinDelayMs;
+
+    @Autowired
+    private NetworkManagementService networkManagementService;
 
     private JChannel channel;
 
@@ -53,6 +61,7 @@ public class FlowClusterService implements Receiver {
 
         System.setProperty("jgroups.tcpping.initial_hosts", seedHosts);
         System.setProperty("jgroups.bind_port", String.valueOf(port));
+        System.setProperty("jgroups.bind_addr", this.bindAddr);
 
         this.channel = new JChannel("jgroups-tcp.xml");
         this.channel.setReceiver(this);
@@ -112,6 +121,11 @@ public class FlowClusterService implements Receiver {
     @Override
     public void viewAccepted(View newView) {
         logger.info("JGroups cluster view updated — members: {}", newView.getMembers());
+        if (newView.getMembers().size() == 1) {
+            this.networkManagementService.onSoloView();
+        } else {
+            this.networkManagementService.onClusterView();
+        }
     }
 
     @Override