+++ services/core/java/core-flow-service/src/main/java/org/ubiquia/core/flow/service/logic/agent/AgentInitializationLogic.java
@@ -6,24 +6,28 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.ubiquia.common.model.ubiquia.entity.AgentEntity;
 import org.ubiquia.common.library.api.config.AgentConfig;
 import org.ubiquia.common.library.api.repository.AgentRepository;
-
+import org.ubiquia.common.model.ubiquia.entity.AgentEntity;
+import org.ubiquia.common.model.ubiquia.entity.NetworkEntity;
+import org.ubiquia.core.flow.repository.NetworkRepository;
 
 /**
  * A service that ensures our Flow instance initializes in the appropriate order.
  */
 @Service
-public class AgentLogic {
+public class AgentInitializationLogic {
 
-    private static final Logger logger = LoggerFactory.getLogger(AgentLogic.class);
+    private static final Logger logger = LoggerFactory.getLogger(AgentInitializationLogic.class);
 
     @Autowired
     private AgentConfig agentConfig;
 
     @Autowired
     private AgentRepository agentRepository;
+
+    @Autowired
+    private NetworkRepository networkRepository;
 
     @Transactional
     public void tryInitializeAgentInDatabase() {
@@ -32,15 +36,20 @@ public class AgentLogic {
 
         var record = this.agentRepository.findById(this.agentConfig.getId());
         if (record.isEmpty()) {
-            logger.info("...no record found in database for agent id: {}..."
-                , this.agentConfig.getId());
+            logger.info("...no record found in database for agent id: {}...",
+                this.agentConfig.getId());
             logger.info("...creating new agent record...");
 
+            var network = new NetworkEntity();
+            network = this.networkRepository.save(network);
+
             var entity = new AgentEntity();
-            entity.setDeployedGraphs(new ArrayList<>());
             entity.setId(this.agentConfig.getId());
+            entity.setDeployedGraphs(new ArrayList<>());
+            entity.setNetwork(network);
             entity = this.agentRepository.save(entity);
-            logger.info("...created ubiquia agent entity with id: {}", entity.getId());
+            logger.info("...created ubiquia agent entity with id: {} in network: {}",
+                entity.getId(), network.getId());
         }
 
         logger.info("...finished initialization.");